login, faq 변경
login, faq 주석 추가

login - 레이아웃,색상 변경, select option 추가, js 추가, 로그인 암호화(주석 처리)
faq - 색상 변경
