+++ src/main/java/com/petcare/web/config/SecurityInitializer.java
@@ -1,0 +1,9 @@
+//
+//  package com.petcare.web.config;
+//  
+//  import org.springframework.security.web.context.AbstractSecurityWebApplicationInitializer;
+//  
+//  public class SecurityInitializer extends AbstractSecurityWebApplicationInitializer {
+//  
+//  }
+
