+++ src/main/java/com/petcare/web/config/ServletConfig.java
@@ -3,8 +3,11 @@ package com.petcare.web.config;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.ComponentScan;
 import org.springframework.context.annotation.Configuration;
+import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+import org.springframework.security.crypto.password.PasswordEncoder;
 import org.springframework.validation.Validator;
 import org.springframework.validation.beanvalidation.LocalValidatorFactoryBean;
+import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.servlet.config.annotation.*;
 import org.springframework.web.servlet.config.annotation.EnableWebMvc;
 import org.springframework.web.servlet.config.annotation.InterceptorRegistry;
@@ -18,6 +21,7 @@ import com.petcare.web.interceptor.UserAuthInterceptor;
 import com.petcare.web.interceptor.HospitalAuthInterceptor;
 // import com.petcare.web.interceptor.HospitalAuthInterceptor;
 import com.petcare.web.interceptor.HospitalLoginInterceptor;
+import com.petcare.web.interceptor.HospitalLogoutInterceptor;
 import com.petcare.web.interceptor.LoginIntercepter;
 import com.petcare.web.interceptor.LogoutInterceptor;
 import com.petcare.web.interceptor.SampleInterceptor;
@@ -45,9 +49,14 @@ public class ServletConfig implements WebMvcConfigurer {
     public void addInterceptors(InterceptorRegistry registry) {
     	
         registry.addInterceptor(new SampleInterceptor());
+        // 일반회원 로그인
         registry.addInterceptor(new LoginIntercepter()).addPathPatterns("/loginPro");
+        // 병원회원 로그인
         registry.addInterceptor(new HospitalLoginInterceptor()).addPathPatterns("/loginPro2");
+        // 로그아웃
         registry.addInterceptor(new LogoutInterceptor()).addPathPatterns("/logout");
+        registry.addInterceptor(new HospitalLogoutInterceptor()).addPathPatterns("/logout");
+        // 일반회원 권한 검사 (로그인 상태 확인), addPathPatterns에 명시된 url에 모두 인터셉터 적용
         registry.addInterceptor(new UserAuthInterceptor()).addPathPatterns("/pet/register")
         .addPathPatterns("/pet/get")
         .addPathPatterns("/pet/modify")
@@ -55,7 +64,8 @@ public class ServletConfig implements WebMvcConfigurer {
         .addPathPatterns("/pet/list")
         .addPathPatterns("/userAppointment/list")
         .addPathPatterns("/userAppointment/modify");
-		registry.addInterceptor(new HospitalAuthInterceptor()).addPathPatterns("/appointment/list")
+        // 병원회원 권한 검사 (로그인 상태 확인), addPathPatterns에 명시된 url에 모두 인터셉터 적용
+        registry.addInterceptor(new HospitalAuthInterceptor()).addPathPatterns("/appointment/list")
 		.addPathPatterns("/appointment/modify");
 		
     }
@@ -92,5 +102,17 @@ public class ServletConfig implements WebMvcConfigurer {
     @Override
     public Validator getValidator() {
         return validator();
+    }
+    
+    /**
+     * 로그인 암호화
+     */
+    
+    // 암호화를 위해 BCryptPasswordEncoder bean 생성
+    // HospitalController에 @Autowire하여 생성 후 @PostMapping("/Join")register에서 회원가입 시 비밀번호 암호화 하여 저장
+    @Bean
+    public BCryptPasswordEncoder passwordEncoder(){
+        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+        return bCryptPasswordEncoder;
     }
 }
