--- src/main/java/com/petcare/web/config/SpringSecurityInitializer.java
@@ -1,11 +1,0 @@
-/*
- * package com.petcare.web.config;
- * 
- * import org.springframework.security.web.context.
- * AbstractSecurityWebApplicationInitializer;
- * 
- * public class SpringSecurityInitializer extends
- * AbstractSecurityWebApplicationInitializer {
- * 
- * }
- */
