+++ src/main/java/com/petcare/web/config/WebConfig.java
@@ -9,7 +9,7 @@ import javax.servlet.Filter;
 public class WebConfig extends AbstractAnnotationConfigDispatcherServletInitializer {
     @Override
     protected Class<?>[] getRootConfigClasses() {
-        return new Class[]{RootConfig.class};
+        return new Class[]{RootConfig.class, /*WebSecurityConfig.class*/ };
     }
 
     @Override
