+++ src/main/java/com/petcare/web/config/WebSecurityConfig.java
@@ -1,37 +1,20 @@
-/*
- * package com.petcare.web.config;
- * 
- * import org.springframework.beans.factory.annotation.Autowired; import
- * org.springframework.context.annotation.Configuration; import
- * org.springframework.security.config.annotation.authentication.builders.
- * AuthenticationManagerBuilder; import
- * org.springframework.security.config.annotation.web.builders.HttpSecurity;
- * import org.springframework.security.config.annotation.web.configuration.
- * EnableWebSecurity; import
- * org.springframework.security.config.annotation.web.configuration.
- * WebSecurityConfigurerAdapter;
- * 
- * @Configuration
- * 
- * @EnableWebSecurity public class WebSecurityConfig extends
- * WebSecurityConfigurerAdapter {
- * 
- * @Autowired private UserAuthenticationProvider authenticationProvider;
- * 
- * @Override protected void configure(HttpSecurity http) throws Exception {
- * http.authorizeRequests() .antMatchers("/css/**", "/js/**",
- * "/img/**").permitAll() .antMatchers("/auth/admin/**").hasRole("ADMIN") //
- * 내부적으로 접두어 "ROLE_"가 붙는다. .antMatchers("/auth/**").hasAnyRole("ADMIN", "USER")
- * // 내부적으로 접두어 "ROLE_"가 붙는다. .anyRequest().authenticated();
- * 
- * http.formLogin() .loginPage("/login") // default
- * .loginProcessingUrl("/loginPro2") .failureUrl("/login?error") // default
- * .defaultSuccessUrl("/index") .usernameParameter("hospitalId")
- * .passwordParameter("hospitalPw") .permitAll();
- * 
- * http.logout() .logoutUrl("/logout") // default .logoutSuccessUrl("/login")
- * .permitAll(); }
- * 
- * @Override protected void configure(AuthenticationManagerBuilder auth) {
- * auth.authenticationProvider(authenticationProvider); } }
- */
+// package com.petcare.web.config;
+//import org.springframework.context.annotation.Bean;
+//import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
+//
+//import org.springframework.security.config.annotation.web.configuration.
+//  EnableWebSecurity; import
+//  org.springframework.security.config.annotation.web.configuration.
+//  WebSecurityConfigurerAdapter;
+//  
+//  @EnableWebSecurity public class WebSecurityConfig extends
+//  WebSecurityConfigurerAdapter {
+//
+//	@Bean
+//	    public BCryptPasswordEncoder passwordEncoder(){
+//	        BCryptPasswordEncoder bCryptPasswordEncoder = new BCryptPasswordEncoder();
+//	        return bCryptPasswordEncoder;
+//	    }
+//  
+//}
+//
