+++ src/main/java/com/petcare/web/interceptor/HospitalAuthInterceptor.java
@@ -23,17 +23,20 @@ public class HospitalAuthInterceptor extends HandlerInterceptorAdapter{
 		}
 	}
 
+	// controller 전에 생성
 	@Override
 	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
 			throws Exception {
 
 		HttpSession httpSession = request.getSession();
-
+		
+		// session에 "hospital" 이라는 속성이 없으면 /login 페이지로 돌아가 다시 로그인 하도록 함
 		if(httpSession.getAttribute("hospital")==null) {
 			saveDestination(request);
 			response.sendRedirect("/login");
 			return false;
 		}
+		// session "hospital"이라는 속성이 있으면 원래 이동하려던 url로 이동
 		return true;
 	}
 }
