+++ src/main/java/com/petcare/web/interceptor/HospitalLoginInterceptor.java
@@ -10,12 +10,15 @@ import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
 import com.petcare.web.domain.Hospital;
 
 public class HospitalLoginInterceptor extends HandlerInterceptorAdapter {
-		// preHandle() : 컨트롤러보다 먼저 수행되는 메서드
+	
+	// 컨트롤러에서 loginPro2에서 view로 넘어갈 때 동작
 	@Override
     public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
         
+		// loginPro2에서 model 객체에 "hospital" 이름으로 저장한 값 가져옴
     	Hospital hospital = (Hospital) modelAndView.getModel().get("hospital");
-
+    	
+    	// hospital에 값이 있으면 session에 "hospital"이라는 속성 추가
         if(hospital != null) {
             HttpSession session = request.getSession();
             session.setAttribute("hospital", hospital);
