+++ src/main/java/com/petcare/web/interceptor/HospitalLogoutInterceptor.java
@@ -1,0 +1,18 @@
+package com.petcare.web.interceptor;
+
+import javax.servlet.http.HttpServletRequest;
+import javax.servlet.http.HttpServletResponse;
+
+import org.springframework.web.servlet.ModelAndView;
+import org.springframework.web.servlet.handler.HandlerInterceptorAdapter;
+
+public class HospitalLogoutInterceptor extends HandlerInterceptorAdapter {
+	
+	@Override
+    public void postHandle(HttpServletRequest request, HttpServletResponse response, Object handler, ModelAndView modelAndView) throws Exception {
+        
+    	request.getSession().invalidate();
+ 
+    }
+
+}
