+++ src/main/java/com/petcare/web/interceptor/UserAuthInterceptor.java
@@ -23,13 +23,14 @@ public class UserAuthInterceptor extends HandlerInterceptorAdapter{
 		}
 	}
 
+	// controller 전에 생성
 	@Override
 	public boolean preHandle(HttpServletRequest request, HttpServletResponse response, Object handler)
 			throws Exception {
 		
-		System.out.println("prehandle 인터셉터");
 		HttpSession httpSession = request.getSession();
 
+		// session에 "user" 이라는 속성이 없으면 /login 페이지로 돌아가 다시 로그인 하도록 함
 		if(httpSession.getAttribute("user")==null) {
 			saveDestination(request);
 			response.sendRedirect("/login");
