--- src/main/java/com/petcare/web/service/UserService.java
@@ -1,14 +1,0 @@
-/*
- * package com.petcare.web.service;
- * 
- * import org.springframework.stereotype.Service;
- * 
- * import com.petcare.web.domain.UserVO;
- * 
- * @Service public class UserService {
- * 
- * public UserVO authenticate(String userId, String password) { return "login";
- * }
- * 
- * }
- */
