+++ src/main/webapp/resources/styles/main_styles.css
@@ -140,7 +140,7 @@ h6::selection
 }
 .form-control
 {
-	color: #db5246;
+	color: #000;
 }
 section
 {
